+++ src/align/sam/sam.jl
@@ -7,6 +7,7 @@ import Automa
 import Automa.RegExp: @re_str
 import Bio: Bio, isfilled
 import Bio.Exceptions: missingerror
+import Bio.RecordHelper: compare_memory, unsafe_parse_decimal
 import BufferedStreams
 
 include("flags.jl")
